train-maker/requirements.txt
ezdxf>=1.1.0
matplotlib>=3.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0

